feat: move zone interfaces to separate file
